requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
fastapi==0.110.0
h11==0.14.0
idna==3.6
psycopg2-binary
pydantic==2.6.3
pydantic_core==2.16.3
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.27.1
pytest==8.1.1
pytest-sqlalchemy-mock==0.1.5
httpx==0.27.0
pre-commit==3.6.2
Faker==24.4.0
python-dotenv==1.0.1